move and merge 1
