+++ fileCode.cs
@@ -4,42 +4,43 @@ using System.IO;
 using System.Windows.Forms;
 using System;
 
-namespace AdressBook.files
+namespace snackShack
 {
     internal static class files
     {
+        #region I/O
         internal static void Write(string filepath, char sep) 
         {
             bool status = File.Exists(filepath); //check if the file exists
-            if (status || Program.debug)
+            if (status || Program.debug) //if file exists, or debug is active, run system
             {
                 try
                 {
-                    using (StreamWriter sw = new StreamWriter(filepath))
+                    using (StreamWriter sw = new StreamWriter(filepath)) //utilize the file path to find the file
                     {
-                        foreach (var c in Program.contacts) //the loop for createing the contents which will be saved
+                        foreach (var c in Program.inventory) //the loop for createing the contents which will be saved
                         {
                             //csv - comma seperated values
-                            //firstname-lastname-email-phone-buisness-notes
-                            string ind = (c.index + 1).ToString();
-                            string line = c.firstname + sep + c.lastname + sep +  //first and last name
-                                c.phone + sep + c.email + sep + c.buisness + sep + //contact information
-                                c.notes + sep + ind; //notes
+                            //name-imagepath-cost-index
+                            string ind = (c.index + 1).ToString(); //indux to string
+                            string cost = c.cost.ToString(); //cost to string
+                            string line = c.name + sep + c.imagePath + sep +  //combine string
+                                cost + sep + ind;
                             sw.WriteLine(line); //write the information to the line
                         }
                     } //streamwriter
                 }
-                catch (Exception ex)
+                catch (Exception ex) //find if exception
                 {
-                    AdressBook.coreCommands.error(Program.preMadeErrorMsg, ex, true); //show error
+                    snackShack.coreCommands.error(constants.preMadeErrorMsg, ex, true); //show error
                 }
             }
-            else
+            else //say if no file
             {
-                AdressBook.coreCommands.error("File Not Found"); //show error
+                snackShack.coreCommands.error("File Not Found"); //show error
             }
         } //the write function
-
+        
         internal static void read(string path, char sep, int min) //the read function
         {
             bool status = File.Exists(path); //check if the file exists
@@ -49,30 +50,27 @@ namespace AdressBook.files
                 {
                     using (StreamReader sr = new StreamReader(path)) //make stringreader
                     {
-                        //csv - comma seperated values
+                        //csv - comma seperated values 
                         //firstname-lastname-email-phone-buisness-notes
                         while (!sr.EndOfStream) //add each line to it one by one
                         {
-                            string contact = sr.ReadLine(); //gets the next line of text from the file
-                            var cont = contact.Split(sep); //splits it by the seperator
-                            if (cont.Length >= min)
+                            string item = sr.ReadLine(); //gets the next line of text from the file
+                            var entry = item.Split(sep); //splits it by the seperator
+                            if (entry.Length >= min)
                             {
-                                Contact c = new Contact //make new constact
+                                inventory c = new inventory //make new constact
                                 {
-                                    firstname = cont[0], //put the realevent input into the releavent field
-                                    lastname = cont[1], //same
-                                    phone = cont[2], //same
-                                    email = cont[3], //same
-                                    buisness = Convert.ToBoolean(cont[4]), //same
-                                    notes = cont[5], //same
-                                    index = Convert.ToInt32(cont[6]) - 1 //more of the same             
+                                    name = entry[0], //put the realevent input into the releavent field
+                                    imagePath = entry[1], //same
+                                    cost = Convert.ToInt32(entry[2]), //same
+                                    index = Convert.ToInt32(entry[3]) -1 //any guesses?
                                 };
-                                Program.contacts.Add(c); //add to list
+                                Program.inventory.Add(c); //add to list
                             }
                             else
                             {
-                             
-                                AdressBook.coreCommands.error("error: below Max Length");
+
+                                coreCommands.error("error: below Max Length"); //show error during read
                             }
                         }
                     }
@@ -89,5 +87,6 @@ namespace AdressBook.files
 
             }
         }
+        #endregion
     } //the code for files
 }
